spam_detector/requirements.txt
pandas
SQLAlchemy
scikit-learn
joblib
fastapi
uvicorn
python-dotenv
psycopg2-binary
